Build rich content helper
