+++ lib/operately/activities/notifications/project_check_in_commented.ex
@@ -5,14 +5,11 @@ defmodule Operately.Activities.Notifications.ProjectCheckInCommented do
     author_id = activity.author_id
     project_id = activity.content["project_id"]
     comment = Operately.Updates.get_comment!(activity.content["comment_id"])
+
     people = Projects.list_notification_subscribers(project_id, exclude: author_id)
+    mentioned =  Operationally.RichContent.lookup_mentioned_people(comment.content["message"])
 
-    mentions = 
-      comment.content["message"]
-      |> ProsemirrorMentions.extract_ids()
-      |> Enum.map(fn id -> Operately.People.get_person!(id) end)
-
-    people = Enum.uniq(people ++ mentions)
+    people = Enum.uniq_by(people ++ mentioned, & &1.id)
 
     notifications = Enum.map(people, fn person ->
       %{
